requirements.txt
celery==5.3.4
Django==4.2.20
django-celery-beat==2.8.1
djangorestframework==3.15.0
dotenv==0.9.9
google-api-python-client==2.113.0
gunicorn==21.2.0
model-bakery==1.20.0
pre-commit==4.0.0
psycopg2-binary==2.9.6
pytest==7.2.1
pytest-django==4.8.0
python-dotenv===1.0.0
redis==4.5.4
virtualenv==20.16.3
